chore: refactor test file locations
