+++ packages/validators/src/test/__fixtures__/componentDefinition.ts
