+++ packages/validators/src/validators/tests/validateComponentDefinition.spec.ts
@@ -1,7 +1,7 @@
-import { validateComponentDefinition } from './validateComponentDefinition';
+import { validateComponentDefinition } from '../validateComponentDefinition';
 import { componentDefinition } from '../../test/__fixtures__/componentDefinition';
 import { describe, it, expect, beforeEach } from 'vitest';
-import { ComponentDefinitionType } from '@/schemas/componentDefinition';
+import { ComponentDefinitionType } from '../../schemas/componentDefinition';
 
 describe('validateComponentDefinition', () => {
   it('should validate the component definition successfully', () => {
